splitting up main into modules
main: core bot functions
common: info that multiple modules need
nb: basic progbot lookups
advance: advance rulebook material or trackers
roll: roll commands
safety: safety commands
