+++ .gitignore
@@ -9,3 +9,4 @@ formattedlog.txt
 log.txt
 prefixes.json
 userlevels.txt
+musicbot.py
