+++ maincommon.py
@@ -1,0 +1,102 @@
+import re
+import koduck
+import pandas as pd
+import settings
+import random
+
+cc_color_dictionary = {"MegaChip": 0xA8E8E8,
+                       "ChitChat": 0xff8000,
+                       "Radical Spin": 0x3f5cff,
+                       "Skateboard Dog": 0xff0000,
+                       "Night Drifters": 0xff0055,
+                       "Mystic Lilies": 0x99004c,
+                       "Leximancy": 0x481f65,
+                       "Underground Broadcast": 0x73ab50,
+                       "New Connections": 0xededed,
+                       "Silicon Skin": 0xf012be,
+                       "The Walls Will Swallow You": 0x734b38,
+                       "MUDSLURP": 0x7687c6,
+                       "Mudslurp": 0x7687c6,
+                       "Tarot": 0xfcf4dc,
+                       "Nyx": 0xa29e14,
+                       "Genso Network": 0xff605d,
+                       "Dark": 0xB088D0,
+                       "Item": 0xffffff,
+                       "Chip": 0xbfbfbf,
+                       "Mystery": 0x000000}
+
+help_categories = {"Lookups": ':mag: **Lookups**',
+                  "Rollers": ':game_die: **Rollers**',
+                  "Helpers": ':thumbsup: **Helpers**',
+                  "Reminders (Base)": ':information_source: **Reminders (Base)**',
+                  "Reminders (Advanced Content)": ':trophy: **Reminders (Advanced Content)**',
+                  "Reminders (Liberation)": ':map: **Reminders (Liberation)**',
+                  "Reminders (DarkChips)": ':smiling_imp: **Reminders (DarkChips)**',
+                  "Safety Tools": ':shield: **Safety Tools**'}
+
+element_df = pd.read_csv(settings.elementfile, sep="\t").fillna('')
+element_category_list = pd.unique(element_df["category"].dropna())
+
+help_df = pd.read_csv(settings.helpfile, sep="\t").fillna('')
+help_df["Response"] = help_df["Response"].str.replace('\\\\n', '\n', regex=True)
+help_cmd_list = [i for i in help_df["Command"] if i]
+help_df["Type"] = help_df["Type"].astype("category")
+help_df["Type"].cat.rename_categories(help_categories, inplace=True)
+help_df["Type"].cat.reorder_categories(list(help_categories.values())+[""], inplace=True)
+
+rulebook_df = pd.read_csv(settings.rulebookfile, sep="\t",  converters = {'Version': str}).fillna('')
+pmc_link = rulebook_df[rulebook_df["Name"] == "Player-Made Repository"]["Link"].iloc[0]
+nyx_link = rulebook_df[rulebook_df["Name"] == "Nyx"]["Link"].iloc[0]
+grid_link = rulebook_df[rulebook_df["Name"] == "Grid-Based Combat"]["Link"].iloc[0]
+rulebook_df = rulebook_df[(rulebook_df["Name"] == "NetBattlers") | (rulebook_df["Name"] == "NetBattlers Advance")]
+# these might start complaining; double check that the labels in the rulebook are exact: captilization and whitespace matter!
+rulebook_df["Type"] = rulebook_df["Type"].astype('category').cat.reorder_categories(["Mobile", "Full Res", "Bonus BattleChips"])
+rulebook_df["Release"] = rulebook_df["Release"].astype('category').cat.reorder_categories(["Beta", "Alpha", "Pre-Alpha", "Version"])
+rulebook_df = rulebook_df.sort_values(["Name", "Release", "Version", "Type"])
+
+playermade_list = ["Genso Network"]
+
+
+def clean_args(args, lowercase=True):
+    if len(args) == 1:
+        args = re.split(r"(?:,|;|\s+)", args[0])
+
+    if lowercase:
+        args = [i.lower().strip() for i in args if i and not i.isspace()]
+    else:
+        args = [i.strip() for i in args if i and not i.isspace()]
+    return args
+
+
+async def send_query_msg(context, return_title, return_msg):
+    return await koduck.sendmessage(context["message"], sendcontent="**%s**\n*%s*" % (return_title, return_msg))
+
+
+async def find_value_in_table(context, df, search_col, search_arg, suppress_notfound=False, alias_message=False, allow_duplicate=False):
+    if "Alias" in df:
+        alias_check = df[
+            df["Alias"].str.contains("(?:^|,|;)\s*%s\s*(?:$|,|;)" % re.escape(search_arg), flags=re.IGNORECASE)]
+        if (alias_check.shape[0] > 1) and (not allow_duplicate):
+            await koduck.sendmessage(context["message"],
+                                     sendcontent="Found more than one match for %s! You should probably let the devs know...")
+            return None
+        if alias_check.shape[0] != 0:
+            search_arg = alias_check.iloc[0][search_col]
+            if alias_message:
+                await koduck.sendmessage(context["message"],
+                                         sendcontent="Found as an alternative name for **%s**!" % search_arg)
+
+    search_results = df[df[search_col].str.contains("\s*^%s\s*$" % re.escape(search_arg), flags=re.IGNORECASE)]
+    if search_results.shape[0] == 0:
+        if not suppress_notfound:
+            await koduck.sendmessage(context["message"],
+                                     sendcontent="I can't find `%s`!" % search_arg)
+        return None
+    elif search_results.shape[0] > 1:
+        if allow_duplicate:
+            return search_results.iloc[random.randrange(0, search_results.shape[0])]
+        else:
+            await koduck.sendmessage(context["message"],
+                                     sendcontent="Found more than one match for %s! You should probably let the devs know..." % search_arg)
+        return None
+    return search_results.iloc[0]
